+++ app/internal/account/repository.go
@@ -21,7 +21,7 @@ func (r *Repository) Create(ctx context.Context, a *Account) error {
 	return err
 }
 
-func (r *Repository) FindByID(ctx context.Context, id int64) (*Account, error) {
+func (r *Repository) FindByID(ctx context.Context, id int) (*Account, error) {
 	a := &Account{}
 	err := r.db.NewSelect().Model(a).Where("account_id = ?", id).Scan(ctx)
 	if err != nil {
@@ -39,7 +39,7 @@ func (r *Repository) FindByEmail(ctx context.Context, email string) (*Account, e
 	return a, nil
 }
 
-func (r *Repository) Delete(ctx context.Context, id int64) error {
+func (r *Repository) Delete(ctx context.Context, id int) error {
 	_, err := r.db.NewDelete().Model(&Account{}).Where("account_id = ?", id).Exec(ctx)
 	return err
 }
