+++ app/internal/license/model.go
@@ -11,7 +11,7 @@ type Licence struct {
 	bun.BaseModel `bun:"table:licence,alias:l" swaggerignore:"true"`
 
 	LicenceID   uuid.UUID `bun:"licence_id,pk,type:uuid,default:gen_random_uuid()" json:"licence_id"`
-	AccountID   int64     `bun:"account_id,notnull"                                json:"account_id"`
+	AccountID   int       `bun:"account_id,notnull"                                json:"account_id"`
 	Expiration  time.Time `bun:"expiration,notnull"                                json:"expiration"`
 	Transaction string    `bun:"transaction"                                       json:"transaction,omitempty"`
 	IsActive    bool      `bun:"is_active,default:true"                            json:"is_active"`
